% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.r
\name{fis-deprecated}
\alias{fis-deprecated}
\alias{fis}
\title{Deprecated class "fis"}
\description{
The \code{fis} class is deprecated and will be removed in future version, use \link{Fis} instead
}
\section{Methods}{

\describe{
\item{\code{infer_output(data, output_index)}}{
WARNING !!! the \code{output_index} is now 1-based indexed in \link{Fis}, was 0-based indexed in \link{fis-deprecated}}
}
}

\seealso{
\link{FisPro-deprecated}
}
\keyword{internal}
