% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisout.R
\docType{class}
\name{FisOut}
\alias{FisOut}
\title{Class "FisOut"}
\description{
The base class of \link{Fis} output (cannot be instantiate)\cr
Use derived classes \link{FisOutCrisp} or \link{FisOutFuzzy}
}
\section{Fields}{

\describe{
\item{\code{name}}{\link{character} vector, The name of the output}
}}

\section{Methods}{

\describe{
\item{\code{range()}}{
\if{latex}{\out{\hfill}}
\describe{
\item{return:}{\link{numeric} vector, The range of the output (min max values)}
}}}
}

