% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.r
\name{FisPro-deprecated}
\alias{FisPro-deprecated}
\title{Deprecated classes in package \pkg{FisPro}.}
\description{
The classes listed below are deprecated and will be removed in future version.
Alternative classes with similar functionality are also mentioned.\cr
Help pages for deprecated functions are available at \code{help("<class>-deprecated")}.
}
\section{Deprecated classes}{

For \link[=fis-deprecated]{fis}, use \link{Fis} instead\cr
For \link[=mf-deprecated]{mf}, use \link{Mf} instead\cr
For \link[=mf_triangular-deprecated]{mf_triangular}, use \link{MfTriangular} instead\cr
For \link[=mf_trapezoidal_inf-deprecated]{mf_trapezoidal_inf}, use \link{MfTrapezoidalInf} instead\cr
For \link[=mf_trapezoidal_sup-deprecated]{mf_trapezoidal_sup}, use \link{MfTrapezoidalSup} instead\cr
For \link[=mf_trapezoidal-deprecated]{mf_trapezoidal}, use \link{MfTrapezoidal} instead\cr
}

\seealso{
\link{fis-deprecated}\cr
\link{mf-deprecated}\cr
\link{mf_triangular-deprecated}\cr
\link{mf_trapezoidal_inf-deprecated}\cr
\link{mf_trapezoidal_sup-deprecated}\cr
\link{mf_trapezoidal-deprecated}
}
\keyword{internal}
