% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.r
\name{mf-deprecated}
\alias{mf-deprecated}
\alias{mf}
\title{Deprecated class "mf"}
\description{
The \code{mf} class is deprecated and will be removed in future version, use \link{Mf} instead
}
\seealso{
\link{FisPro-deprecated}
}
\keyword{internal}
