% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf.R
\docType{class}
\name{Mf}
\alias{Mf}
\title{Class "Mf"}
\description{
The base class of all "membership function" classes (cannot be instantiate)\cr
Use derived classes \link{MfTriangular}, \link{MfTrapezoidal}, \link{MfTrapezoidalInf} or \link{MfTrapezoidalSup}
}
\section{Fields}{

\describe{
\item{\code{label}}{\link{character} vector, The label of the membership function}
}}

\section{Methods}{

\describe{
\item{\code{degree(value)}}{
Get the membership degree
\if{latex}{\out{\hfill}}
\describe{
\item{argument: \code{value}}{\link{numeric} value to compute the membership degree}
\item{return:}{\link{numeric} value}
}}}
}

\seealso{
{
\href{https://www.fispro.org/documentation/en/inline-help/node39.html}{Fuzzy Logic Elementary Glossary}
}
}
