% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.r
\name{mf_trapezoidal-deprecated}
\alias{mf_trapezoidal-deprecated}
\alias{mf_trapezoidal}
\title{Deprecated class "mf_trapezoidal"}
\description{
The \code{mf_trapezoidal} class is deprecated and will be removed in future version, use \link{MfTrapezoidal} instead
}
\seealso{
\link{FisPro-deprecated}
}
\keyword{internal}
