% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\docType{class}
\name{mf_trapezoidal}
\alias{mf_trapezoidal}
\title{mf_trapezoidal class}
\description{
Class to build trapezoidal MF.
}
\section{Inherits}{

mf_trapezoidal class inherits all fields and methods of \link{mf} class.
}

\section{Methods}{

\describe{
\item{\code{constructor(lower_support, lower_kernel, upper_kernel, upper_support)}}{argument: \code{lower_support} numerical lower value of support. \cr argument: \code{lower_kernel} numerical lower value of kernel. \cr argument: \code{upper_kernel} numerical upper value of kernel. \cr argument: \code{upper_support} numerical upper value of support. \cr return: an object of mf_trapezoidal.}
}
}

\examples{
mf <- new(mf_trapezoidal, 0, 1, 2, 3)
mf$degree(0.5)

}
