% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.r
\name{mf_trapezoidal_inf-deprecated}
\alias{mf_trapezoidal_inf-deprecated}
\alias{mf_trapezoidal_inf}
\title{Deprecated class "mf_trapezoidal_inf"}
\description{
The \code{mf_trapezoidal_inf} class is deprecated and will be removed in future version, use \link{MfTrapezoidalInf} instead
}
\seealso{
\link{FisPro-deprecated}
}
\keyword{internal}
