% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\docType{class}
\name{mf_trapezoidal_inf}
\alias{mf_trapezoidal_inf}
\title{mf_trapezoidal_inf class}
\description{
Class to build trapezoidal inf MF.
}
\section{Inherits}{

mf_trapezoidal_inf class inherits all fields and methods of \link{mf} class.
}

\section{Methods}{

\describe{
\item{\code{constructor(upper_kernel, upper_support)}}{argument: \code{upper_kernel} numerical upper value of kernel. \cr argument: \code{upper_support} numerical upper value of support. \cr return: an object of mf_trapezoidal_inf.}
}
}

\examples{
mf <- new(mf_trapezoidal_inf, 0, 1)
mf$degree(0.5)

}
