% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.r
\name{mf_trapezoidal_sup-deprecated}
\alias{mf_trapezoidal_sup-deprecated}
\alias{mf_trapezoidal_sup}
\title{Deprecated class "mf_trapezoidal_sup"}
\description{
The \code{mf_trapezoidal_sup} class is deprecated and will be removed in future version, use \link{MfTrapezoidalSup} instead
}
\seealso{
\link{FisPro-deprecated}
}
\keyword{internal}
