% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\docType{class}
\name{mf_trapezoidal_sup}
\alias{mf_trapezoidal_sup}
\title{mf_trapezoidal_sup class}
\description{
Class to build trapezoidal sup MF.
}
\section{Inherits}{

mf_trapezoidal_sup class inherits all fields and methods of \link{mf} class.
}

\section{Methods}{

\describe{
\item{\code{constructor(lower_support, lower_kernel)}}{argument: \code{lower_support} numerical lower value of support. \cr argument: \code{lower_kernel} numerical lower value of kernel. \cr return: an object of mf_trapezoidal_sup.}
}
}

\examples{
mf <- new(mf_trapezoidal_sup, 0, 1)
mf$degree(0.5)

}
