% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.r
\name{mf_triangular-deprecated}
\alias{mf_triangular-deprecated}
\alias{mf_triangular}
\title{Deprecated class "mf_triangular"}
\description{
The \code{mf_triangular} class is deprecated and will be removed in future version, use \link{MfTriangular} instead
}
\seealso{
\link{FisPro-deprecated}
}
\keyword{internal}
