% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\docType{class}
\name{mf_triangular}
\alias{mf_triangular}
\title{mf_triangular class}
\description{
Class to build triangular MF.
}
\section{Inherits}{

mf_triangular class inherits all fields and methods of \link{mf} class.
}

\section{Methods}{

\describe{
\item{\code{constructor(lower_support, kernel, upper_support)}}{argument: \code{lower_support} numerical lower value of support. \cr argument: \code{kernel} numerical value of kernel. \cr argument: \code{upper_support} numerical upper value of support. \cr return: an object of mf_triangular.}
}
}

\examples{
mf <- new(mf_triangular, 0, 1, 2)
mf$degree(0.5)

}
