% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf.R
\docType{class}
\name{MfTrapezoidal}
\alias{MfTrapezoidal}
\title{Class "MfTrapezoidal"}
\description{
Class to manage a trapezoidal membership function
}
\section{Inherits}{

MfTrapezoidal class inherits all fields and methods of \link{Mf} class
}

\section{Constructors}{

\describe{
\item{\code{MfTrapezoidal(lower_support, lower_kernel, upper_kernel, upper_support)}}{
\if{latex}{\out{\hfill}}
\describe{
\item{argument: \code{lower_support}}{\link{numeric} lower value of support}
\item{argument: \code{lower_kernel}}{\link{numeric} lower value of kernel}
\item{argument: \code{upper_kernel}}{\link{numeric} upper value of kernel}
\item{argument: \code{upper_support}}{\link{numeric} upper value of support}
\item{return:}{\link{MfTrapezoidal} object}
}}}
}

\examples{
mf <- NewMfTrapezoidal(0, 1, 2, 3)
mf$degree(0.5)
}
\seealso{
\link{NewMfTrapezoidal}
}
