% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf.R
\docType{class}
\name{MfTrapezoidalInf}
\alias{MfTrapezoidalInf}
\title{Class "MfTrapezoidalInf"}
\description{
Class to manage a trapezoidal inf membership function
}
\section{Inherits}{

MfTrapezoidalInf class inherits all fields and methods of \link{Mf} class
}

\section{Constructors}{

\describe{
\item{\code{MfTrapezoidalInf(upper_kernel, upper_support)}}{
\if{latex}{\out{\hfill}}
\describe{
\item{argument: \code{upper_kernel}}{\link{numeric} upper value of kernel}
\item{argument: \code{upper_support}}{\link{numeric} upper value of support}
\item{return:}{\link{MfTrapezoidalInf} object}
}}}
}

\examples{
mf <- NewMfTrapezoidalInf(0, 1)
mf$degree(0.5)
}
\seealso{
\link{NewMfTrapezoidalInf}
}
