% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf.R
\docType{class}
\name{MfTrapezoidalSup}
\alias{MfTrapezoidalSup}
\title{Class "MfTrapezoidalSup"}
\description{
Class to manage a trapezoidal sup membership function
}
\section{Inherits}{

MfTrapezoidalSup class inherits all fields and methods of \link{Mf} class
}

\section{Constructors}{

\describe{
\item{\code{MfTrapezoidalSup(lower_support, lower_kernel)}}{
\if{latex}{\out{\hfill}}
\describe{
\item{argument: \code{lower_support}}{\link{numeric} lower value of support}
\item{argument: \code{lower_kernel}}{\link{numeric} lower value of kernel}
\item{return:}{\link{MfTrapezoidalSup} object}
}}}
}

\examples{
mf <- NewMfTrapezoidalSup(0, 1)
mf$degree(0.5)
}
\seealso{
\link{NewMfTrapezoidalSup}
}
