% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf.R
\docType{class}
\name{MfTriangular}
\alias{MfTriangular}
\title{Class "MfTriangular"}
\description{
Class to manage a triangular membership function
}
\section{Inherits}{

\code{MfTriangular} class inherits all fields and methods of \link{Mf} class
}

\section{Constructors}{

\describe{
\item{\code{MfTriangular(lower_support, kernel, upper_support)}}{
\if{latex}{\out{\hfill}}
\describe{
\item{argument: \code{lower_support}}{\link{numeric} lower value of support}
\item{argument: \code{kernel}}{\link{numeric} value of kernel}
\item{argument: \code{upper_support}}{\link{numeric} upper value of support}
\item{return:}{\link{MfTriangular} object}
}}}
}

\examples{
mf <- NewMfTriangular(0, 1, 2)
mf$degree(0.5)
}
\seealso{
\link{NewMfTriangular}
}
