% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fis.R
\docType{methods}
\name{NewFis}
\alias{NewFis}
\title{Create object of class "Fis"}
\usage{
NewFis(...)
}
\arguments{
\item{...}{arguments of \link{Fis} constructor}
}
\value{
\link{Fis} object
}
\description{
Function to create object of class \link{Fis}
}
