% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisin.R
\docType{methods}
\name{NewFisIn}
\alias{NewFisIn}
\title{Create object of class "FisIn"}
\usage{
NewFisIn(...)
}
\arguments{
\item{...}{arguments of \link{FisIn} constructor}
}
\value{
\link{FisIn} object
}
\description{
Function to create object of class \link{FisIn}
}
