% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisout.R
\docType{methods}
\name{NewFisOutCrisp}
\alias{NewFisOutCrisp}
\title{Create object of class "FisOutCrisp"}
\usage{
NewFisOutCrisp(...)
}
\arguments{
\item{...}{arguments of \link{FisOutCrisp} constructor}
}
\value{
\link{FisOutCrisp} object
}
\description{
Function to create object of class \link{FisOutCrisp}
}
