% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisout.R
\docType{methods}
\name{NewFisOutFuzzy}
\alias{NewFisOutFuzzy}
\title{Create object of class "FisOutFuzzy"}
\usage{
NewFisOutFuzzy(...)
}
\arguments{
\item{...}{arguments of \link{FisOutFuzzy} constructor}
}
\value{
\link{FisOutFuzzy} object
}
\description{
Function to create object of class \link{FisOutFuzzy}
}
