% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf.R
\docType{methods}
\name{NewMfTrapezoidal}
\alias{NewMfTrapezoidal}
\title{Create object of class "MfTrapezoidal"}
\usage{
NewMfTrapezoidal(...)
}
\arguments{
\item{...}{arguments of \link{MfTrapezoidal} constructor}
}
\value{
\link{MfTrapezoidal} object
}
\description{
Function to create object of class \link{MfTrapezoidal}
}
