% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf.R
\docType{methods}
\name{NewMfTrapezoidalInf}
\alias{NewMfTrapezoidalInf}
\title{Create object of class "MfTrapezoidalInf"}
\usage{
NewMfTrapezoidalInf(...)
}
\arguments{
\item{...}{arguments of \link{MfTrapezoidalInf} constructor}
}
\value{
\link{MfTrapezoidalInf} object
}
\description{
Function to create object of class \link{MfTrapezoidalInf}
}
