% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf.R
\docType{methods}
\name{NewMfTrapezoidalSup}
\alias{NewMfTrapezoidalSup}
\title{Create object of class "MfTrapezoidalSup"}
\usage{
NewMfTrapezoidalSup(...)
}
\arguments{
\item{...}{arguments of \link{MfTrapezoidalSup} constructor}
}
\value{
\link{MfTrapezoidalSup} object
}
\description{
Function to create object of class \link{MfTrapezoidalSup}
}
