% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf.R
\docType{methods}
\name{NewMfTriangular}
\alias{NewMfTriangular}
\title{Create object of class "MfTriangular"}
\usage{
NewMfTriangular(...)
}
\arguments{
\item{...}{arguments of \link{MfTriangular} constructor}
}
\value{
\link{MfTriangular} object
}
\description{
Function to create object of class \link{MfTriangular}
}
