% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule.R
\docType{methods}
\name{NewRule}
\alias{NewRule}
\title{Create object of class "Rule"}
\usage{
NewRule(...)
}
\arguments{
\item{...}{arguments of \link{Rule} constructor}
}
\value{
\link{Rule} object
}
\description{
Function to create object of class \link{Rule}
}
