% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\docType{class}
\name{mf}
\alias{mf}
\title{mf class}
\description{
Base class of all MF classes. This class cannot be instantiate. Use derived classes \link{mf_triangular}, \link{mf_trapezoidal}, \link{mf_trapezoidal_inf} or \link{mf_trapezoidal_sup}
}
\section{Fields}{

\describe{
\item{\code{name}}{the name of the mf.}
}}

\section{Methods}{

\describe{
\item{\code{degree(value)}}{compute the membership degree of a numerical value. \cr argument: \code{value} numerical value to compute the membership degree. \cr return: the membership degree.}
}
}

\references{
{
  \href{https://www.fispro.org/documentation/en/inline-help/node39.html}{Fuzzy Logic Elementary Glossary}
}
}
